Refactor mute handling in test scripts and update paths for consistency
- Replaced references to `get_muted_tests.py` with `test_report_mute.py` in various scripts to align with the new structure.
- Deleted the obsolete `get_muted_tests.py` file to streamline the codebase.
- Updated import statements in `junit-postprocess.py` and `mute_check.py` to reflect the new module organization, enhancing maintainability.
